requirements.txt
# Core dependencies for news-podcast-agent
# Based on pyproject.toml dependencies

# Google Cloud and AI Platform
google-adk~=1.6.0
opentelemetry-exporter-gcp-trace~=1.9.0
google-cloud-logging~=3.11.4
google-cloud-aiplatform[evaluation,agent-engines]~=1.103.0

# Environment and configuration
python-dotenv~=1.0.0

# Web framework
fastapi~=0.104.0
uvicorn[standard]~=0.24.0
pydantic~=2.5.0
python-multipart~=0.0.6

# Additional dependencies for news API
newsapi-python~=0.2.6

# Audio processing (if needed)
google-cloud-texttospeech~=2.16.0

# Development dependencies (optional)
# pytest>=8.3.4
# pytest-asyncio>=0.23.8
# nest-asyncio>=1.6.0
# ruff>=0.4.6
# mypy~=1.15.0
# codespell~=2.2.0